requirements.txt
ccxt==4.*
pandas>=2.0.0
pandas-ta>=0.3.14b0
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pyupbit>=0.2.33
